moved SummaryList into its own folder, separated component from container
