+++ src/components/SummaryList/index.js
@@ -1,0 +1,1 @@
+export { default } from './SummaryList.container';
